pyproject.toml

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eff-mcp"
version = "0.1.0"
description = "Ethics Filter Framework MCP server"
authors = [
	{ name = "Vajo Sekulic" }
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"openai>=1.0.0",
	"fastmcp>=0.1.0",
	"pydantic>=2.0.0",
	"python-dotenv>=1.0.0"
]

[project.optional-dependencies]
rag = ["supabase>=2.0.0"]
dev = ["pytest>=7.0"]

[project.scripts]
eff-mcp = "eff.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["eff*"]

[tool.setuptools.package-data]
"eff.resources" = ["*.json", "*.md"]

[tool.pytest.ini_options]
testpaths = ["tests"]
